Adopt ArgumentParser and AbsolutePath in the prebuilt build script.
